+++ .aws-sam/build.toml
@@ -1,0 +1,15 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions.3b5b3f0a-eca5-4cd7-911a-15fe01dcb036]
+codeuri = "/Users/yoshi/aws_sam/youtube-line-notification-app/line-notification"
+runtime = "provided.al2023"
+architecture = "x86_64"
+handler = "line-notification"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["LineNotificationFunction"]
+
+[function_build_definitions.3b5b3f0a-eca5-4cd7-911a-15fe01dcb036.metadata]
+BuildMethod = "go1.x"
+
+[layer_build_definitions]
