--- hello-world/go.mod
@@ -1,7 +1,0 @@
-require github.com/aws/aws-lambda-go v1.36.1
-
-replace gopkg.in/yaml.v2 => gopkg.in/yaml.v2 v2.2.8
-
-module hello-world
-
-go 1.16
