--- hello-world/main_test.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"testing"
-)
-
-func TestHandler(t *testing.T) {
-
-}
