+++ line-notification/go.mod
@@ -1,0 +1,7 @@
+require github.com/aws/aws-lambda-go v1.36.1
+
+replace gopkg.in/yaml.v2 => gopkg.in/yaml.v2 v2.2.8
+
+module line-notification
+
+go 1.16
