+++ line-notification/main_test.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestHandler(t *testing.T) {
+
+}
