+++ template.yaml
@@ -11,13 +11,13 @@ Globals:
     MemorySize: 128
 
 Resources:
-  HelloWorldFunction:
+  LineNotificationFunction:
     Type: AWS::Serverless::Function
     Metadata:
       BuildMethod: go1.x
     Properties:
-      CodeUri: hello-world/
-      Handler: hello-world
+      CodeUri: line-notification/
+      Handler: line-notification
       Runtime: provided.al2023
       Architectures:
         - x86_64
